+++ discovery/config/config.go
@@ -11,9 +11,30 @@ const (
 )
 
 type Config interface {
+	// Endpoint is a required starting endpoint for connect
 	Endpoint() string
+
+	// Database is a required database name.
 	Database() string
+
+	// Secure is an flag for secure connection
 	Secure() bool
+
+	// OperationTimeout is the maximum amount of time a YDB server will process
+	// an operation. After timeout exceeds YDB will try to cancel operation and
+	// regardless of the cancellation appropriate error will be returned to
+	// the client.
+	// If OperationTimeout is zero then no timeout is used.
+	OperationTimeout() time.Duration
+
+	// OperationCancelAfter is the maximum amount of time a YDB server will process an
+	// operation. After timeout exceeds YDB will try to cancel operation and if
+	// it succeeds appropriate error will be returned to the client; otherwise
+	// processing will be continued.
+	// If OperationCancelAfter is zero then no timeout is used.
+	OperationCancelAfter() time.Duration
+
+	// Trace defines trace over discovery client calls
 	Trace() trace.Discovery
 
 	// Interval is the frequency of background tasks of ydb endpoints discovery.
@@ -26,8 +47,20 @@ type config struct {
 	endpoint string
 	database string
 	secure   bool
+
+	operationTimeout     time.Duration
+	operationCancelAfter time.Duration
+
 	interval time.Duration
 	trace    trace.Discovery
+}
+
+func (c *config) OperationTimeout() time.Duration {
+	return c.operationTimeout
+}
+
+func (c *config) OperationCancelAfter() time.Duration {
+	return c.operationCancelAfter
 }
 
 func (c *config) Interval() time.Duration {
@@ -73,6 +106,18 @@ func WithSecure(ssl bool) Option {
 func WithTrace(trace trace.Discovery) Option {
 	return func(c *config) {
 		c.trace = trace
+	}
+}
+
+func WithOperationTimeout(operationTimeout time.Duration) Option {
+	return func(c *config) {
+		c.operationTimeout = operationTimeout
+	}
+}
+
+func WithOperationCancelAfter(operationCancelAfter time.Duration) Option {
+	return func(c *config) {
+		c.operationCancelAfter = operationCancelAfter
 	}
 }