+++ test/ratelimiter_test.go
@@ -31,8 +31,6 @@ func TestRatelimiter(t *testing.T) {
 		ydb.WithConnectionString(os.Getenv("YDB_CONNECTION_STRING")),
 		ydb.WithAnonymousCredentials(),
 		ydb.With(
-			config.WithRequestTimeout(time.Second*2),
-			config.WithStreamTimeout(time.Second*2),
 			config.WithOperationTimeout(time.Second*2),
 			config.WithOperationCancelAfter(time.Second*2),
 		),
@@ -95,7 +93,7 @@ func TestRatelimiter(t *testing.T) {
 		// cleanup resource
 		err = db.Ratelimiter().DropResource(ctx, testCoordinationNodePath, testResource)
 		if err != nil {
-			t.Fatal("Cannot drop resource")
+			t.Fatalf("Cannot drop resource: %v", err)
 		}
 	}()
 	// describe resource